delete: HTTPClientAdaptor
